requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
supabase==2.3.0
python-dotenv==1.0.0
transformers==4.36.2
torch==2.1.2
Pillow==10.1.0
tqdm==4.66.1
pandas==2.1.4
asyncio-mqtt==0.16.1
aiohttp==3.9.1
tenacity==8.2.3
psycopg2-binary==2.9.9
